# openapi.json
{"openapi":"3.1.0","info":{"title":"Flock Energy API","description":"Clean REST API wrapper around the Urja Meter Ops portal.","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}},"/api/v1/meters":{"get":{"summary":"List Meters","operationId":"list_meters_api_v1_meters_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","default":"","title":"Q"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeterListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}":{"get":{"summary":"Get Meter","operationId":"get_meter_api_v1_meters__meter_id__get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeterDetailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}/location":{"get":{"summary":"Get Meter Location","operationId":"get_meter_location_api_v1_meters__meter_id__location_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}/consumption":{"get":{"summary":"Get Meter Consumption","operationId":"get_meter_consumption_api_v1_meters__meter_id__consumption_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsumptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ConsumptionReading":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"kwh":{"type":"number","title":"Kwh"},"kvah":{"type":"number","title":"Kvah"},"voltage_r":{"type":"number","title":"Voltage R"}},"type":"object","required":["timestamp","kwh","kvah","voltage_r"],"title":"ConsumptionReading"},"ConsumptionResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"readings":{"items":{"$ref":"#/components/schemas/ConsumptionReading"},"type":"array","title":"Readings"}},"type":"object","required":["meter_id","readings"],"title":"ConsumptionResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Location":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"}},"type":"object","required":["meter_id","latitude","longitude"],"title":"Location"},"Meter":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"serial_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Serial Number"},"make":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Make"},"phase":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phase"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"dt_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dt Code"}},"type":"object","required":["meter_id"],"title":"Meter"},"MeterDetailResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"details":{"$ref":"#/components/schemas/MeterDetails"},"hierarchy":{"$ref":"#/components/schemas/MeterHierarchy"}},"type":"object","required":["meter_id","details","hierarchy"],"title":"MeterDetailResponse"},"MeterDetails":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"serial_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Serial Number"},"make":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Make"},"phase_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phase Type"},"installation_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Installation Status"},"installation_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Installation Type"}},"type":"object","required":["meter_id"],"title":"MeterDetails"},"MeterHierarchy":{"properties":{"zone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Zone"},"circle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Circle"},"division":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Division"},"subdivision":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subdivision"},"substation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Substation"},"feeder":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Feeder"},"dt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dt"}},"type":"object","title":"MeterHierarchy"},"MeterListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Meter"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"}},"type":"object","required":["items","total","page","page_size"],"title":"MeterListResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}